backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
motor==3.6.0
beanie==1.27.0
pydantic[email]==2.10.4
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
resend==2.6.0
google-genai==0.3.0
torch==2.5.1
transformers==4.46.3
joblib==1.4.2
numpy==1.26.4
xgboost==2.1.2
textblob==0.18.0.post0
sentence-transformers==3.3.1
deep-translator==1.11.4
scikit-learn==1.5.2
certifi==2024.8.30
protobuf==5.28.3
